requirements.txt
nextcord
python-dotenv
pandas
requests
bs4
unidecode